added transaction db

+++ models/positions.model.go
@@ -7,7 +7,7 @@ import (
 type Positions struct {
 	gorm.Model
 	Name      string
-	Salary    uint
+	Salary    uint64
 }
